requirements.txt
nltk
streamlit
st-annotated-text
pandas